requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
httpx==0.25.2
python-dotenv==1.0.0
chromadb==0.4.22
sentence-transformers==2.2.2
numpy==1.24.3
facebook-scraper>=0.2.60
beautifulsoup4>=4.14.0
playwright>=1.40.0
huggingface_hub>=0.20.0

